4.21 Change the delete blog operation so that a blog can be deleted only by the user who added it

+++ part4/blogList/controllers/blogs.js
@@ -40,7 +40,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === decodedToken.id.toString() ){
+    await Blog.findByIdAndDelete(request.params.id)
+  }else{
+    return response.status(403).json({ error: 'only the creator can delete this blog' })
+  }
+
   response.status(204).end()
 })
